Added cover platform support
