+++ boschshc/__init__.py
@@ -40,7 +40,7 @@ CONFIG_SCHEMA = vol.Schema(
     extra=vol.ALLOW_EXTRA,
 )
 
-PLATFORMS = ["binary_sensor"]
+PLATFORMS = ["binary_sensor", "cover"]
 
 _LOGGER = logging.getLogger(__name__)
